+++ src/Guard.php
@@ -11,7 +11,7 @@ use Illuminate\Support\Collection;
 class Guard
 {
     /**
-     * return collection of (guard_name) property if exist on class or object
+     * return collection of (guardName) property if exist on class or object
      * otherwise will return collection of guards names that exists in config/auth.php.
      *
      * @param $model
@@ -25,12 +25,12 @@ class Guard
         $class = null;
 
         if (\is_object($model)) {
-            $guardName = $model->guard_name ?? null;
+            $guardName = $model->guardName ?? null;
         }
 
         if ($guardName === null) {
             $class = \is_object($model) ? \get_class($model) : $model;
-            $guardName = (new \ReflectionClass($class))->getDefaultProperties()['guard_name'] ?? null;
+            $guardName = (new \ReflectionClass($class))->getDefaultProperties()['guardName'] ?? null;
         }
 
         if ($guardName) {